WIP: Interaction outline
I created a custom render feature that reads the player’s interaction handler, and draws an inverse hull outline around it. The system displays differently coloured outlines for if an interactable is enabled or disabled, or if it’s just a physics prop.
This system works fairly well, but the outline itself doesn’t look great. The shape can be wonky and inconsistent, and parts of the outline will still clip in front of the object. That’s all minor cosmetic stuff though, that I can fix later.