WIP: Object heat and thermal vision for player
WIP: Object heat and thermal vision for player
If the player is spraying smoke everywhere, that will impede their own vision as well as the enemy’s, which can make combat somewhat annoying. To remedy this, I’m adding thermal vision. This effect relies on a few main systems:
•	ObjectHeat, a MonoBehaviour that attaches to objects in the scene, and contains a heat value in degrees This class also includes a static ambient temperature and a temperature range.
•	A thermal vision shader that shifts through a gradient and opacity based on a 0-1 heat value.
•	A second forward renderer containing a custom ‘RenderThermalVision’ render feature. This first renders an overlay on all scene objects to reflect the ambient heat value, then renders over specific renderers attached to ObjectHeat scripts, to display unique temperature values. It obtains the correct 0-1 value by performing an inverse lerp on the desired temperature, and the minimum and maximum values of the temperature range. Opacity is also set based on if the renderer is on the smoke layer.
•	A ‘PlayerThermalVision’ script that switches the player cameras between this forward renderer and the standard one.
This system is a WIP and there are several bugs to fix, as well as new cosmetic features to implement. For example, transparency doesn’t actually show properly, and I need to implement proper animations and controls for switching to and from thermal vision. But I’m extremely pleased with how this system is turning out.
I also intend to include thermal vision as a feature for some enemy types, but this will be much easier to implement as it won’t require any shader shenanigans.