MiscFunctions.GetImmedateComponentsInChildren()
This function obtains child components of type A from a parent component of type B, whose closest parent component of type B is the specified target. This function is for obtaining child components of entities, while ignoring components that belong to other entities that are children of the first entity.
I forgot to include this script in a previous commit.