Fixed bug where thermal vision would not show proper object depth
This was due to the thermal vision material being transparent to account for smoke. Problem is that transparent objects don't render their own depth passes. So to fix this, I just added an extra pass of an extremely simple opaque material, to save me the trouble of having a separate opaque version of my thermal vision material.