State machine system now inherently supports asynchronous functions for starting and ending a state
Now a state can include not only asynchronous/delayed functions that run automatically when the state is entered, but can also include exit functions, that the controller will wait to run before switching to the next state. This system allows for far more flexibility in switching between different player and enemy actions.