WIP: offhand attack handler
This new class will allow the player to switch between multiple offhand abilities, and when the appropriate input is pressed the player-character will automatically put away their weapon (if necessary) to perform it. This code will also be used for things such as picking up and throwing physics objects, and maybe even contextual interactions in general.