Map screen WIP: map menu, exploration fill and map shader
The player can now press a button to open a map menu, which renders a mesh for the current level. The player's progress moving about the level is saved to a 3D texture, which is then drawn from to only render the parts of the map mesh that the player has been to (unless the auto-map bool is set to true).