Updated player input controls to match offhand attack being more important
Offhand attacks are now performed using the right mouse button or left trigger, and aiming down sights is now toggled with the middle mouse button or clicking in the right analog stick. Kicking controls the same on keyboard, but on a controller is now done using the left shoulder button.