Thrown object trajectory can now be visualised before throwing, using new TrajectoryRenderer class
The player can now hold down the throw button to show the trajectory of a thrown object. This change also means that if the player is holding a two-handed weapon (which is automatically put away before throwing), they are still able to avoid throwing until they've lined up their aim perfectly.