Offhand ability cancelling (and a whole heap of bugfixes for throwing objects)
All held offhand abilities can now be cancelled by pressing the 'interact' button or switching weapons, not just carrying and throwing physics props.
This change came with having to fix a whole lot of bugs caused by throwing objects no longer being an instant action, and also cancellable. I had to create a function in ThrowHandler that automatically drops an item and disables the arc renderer, if the throw option is cancelled instead of completed. I also had to add code in ThrowObject to still keep the thrown object instead of dropping it, since it wasn't meant to be spawned in otherwise. I also had to heavily standardise the code for cancelling attacks and move it all into OffhandAttackHandler.