Fixed bug where decals wouldn't display correctly on Linux
I changed the decal feature's technique to 'screen space', then set 'use rendering layers' to true to ensure it didn't mess up the outline effect on Windows. Not sure why it affects the outlines, or why  the outlines were messing up in the same way that it does on Linux, or why the dodgy outlines on Linux are the same before and after these changes.