Fixed bug where AI characters couldn't climb stairs
This was caused by the navmesh velocity not being enough to counteract the force of gravity. I fixed this by reusing the player's grounding code to check if the enemy is on solid ground, and disabling gravity while this is the case. If the enemy becomes airborne, gravity will automatically kick in and make them start falling again.