Refactored ADS code to fix bugs preventing its use
The previous ADS code was included within GunADS itself, which caused bugs when a gun was equipped on an enemy (since enemies don’t use those mechanics and they’re reliant on player-specific data). To fix this, I added a whole heap of checks to disable the function, but this ended up disabling ADS for other weird reasons. While troubleshooting this, I noticed the code was quite messy, so I decided to totally refactor it.
I moved the actual ADS code (transitioning animations, changing weapon accuracy, zooming in the camera, etc.) to a new class called ADSHandler, which is located on the player and referenced by WeaponHandler. Whenever the player switches to a weapon with ADS data, it’s loaded into ADSHandler so the appropriate changes can be applied when the player enables/disables the ability.