Refactored button prompts to now switch between single and multiple prompts
This change was necessary for the map screen directional inputs, which can be either a composite of distinct buttons or a single directional input. E.g. for zooming, which uses the scroll wheel on mouse and keyboard but the left and right triggers on a gamepad.
The new code is somewhat similar to the deprecated MultiAxisButtonPrompt, except if an available single input is present, it will just display that, hide the other prompt graphics and break the loop of searching through the available prompts. The tutorial prompt prefab now uses a HorizontalLayoutGroup and ContentSizeFitter to ensure the entire prompt is only as big as it needs to be for the inputs shown.
With these changes, I’ve also been able to delete MultiAxisButtonPrompt and GUIButtonPromptBase, and shrink GUIButtonPrompt down to being a standalone class again, that just handles showing individual bindngs for keys. Overall, I’m quite pleased with these changes, and how quickly I was able to implement them.