Fixed bug where object throwing GUI wouldn't go away when dropping object
I fixed this by assigning it directly to the ThrowHandler rather than the PropCarryingHandler. This also means the GUI appears when preparing to throw a smoke grenade, which is something I wanted to add anyway.
In this commit, I also renamed the button input names, as I'd accidentally deleted them when reworking the GUI button prompts.