Moved most of my shader code from this project into a separate package
This ensures I can show off said shader code more effectively on my portfolio, and reuse it in other projects when I need to. The thermal vision code is still here because it's too unique to this game's mechanics, and the map fill code is also still here because I haven't figured out what parts should be moved into the shader package.