Condensing essential level scripts into a single prefab, to expedite making new levels
The new 'Level Core' prefab contains the AI grid points, objective handler and navmeshes already in place, so I don't have to add them manually. I also took the cached terrain grid code from TerrainMap and put it into its own self-contained class (TerrainGrid) as it will need to be used for many other functions and is not based on data exclusive to the player. TerrainMap still works on filling in the map as the player moves about the level.