Terrain grid is now handled by separate class, with chunk support
The code for creating the world grid as a BoundsInt is now handled by a new TerrainGrid class, since it's a global value for the current level and will be used for things besides the player map. I also added support for chunks - based on a chunk size, the world grid will now be expanded to support a whole number of chunks. This will be useful for the new smoke system I'm making, as splitting the simulation into chunks will allow me to only simulate sections with active smoke.