Smoke grid now uses chunks, ensuring only parts with smoke inside them need to be simulated
SmokeChunks handle the individual simulation, and if smoke goes outside the bounds it will automatically check for a neighbouring chunk. The SimulatedSmokeGrid class stores the base stats like timestep and max density, and iterates through each chunk to ensure everything's simulated at a consistent speed and in the correct order. E.g. ensuring that one stage of a simulation is complete in all chunks before moving onto the next.