Swapped around controller inputs for crouching and sprinting
Pressing a face button is much quicker to do, and easier to do while pushing the stick in a specific direction, making it better suited for needing to quickly sprint away from danger. Whereas crouching is a slower, more deliberate action, so it makes sense to relegate it to a button that's annoying to press regularly.