Optimised smoke cloud rendering using lower-poly sphere mesh
The amount of polygons on the default Unity sphere is complete overkill for the level of detail I actually need (which is to say barely any). The natural smoothness is already broken up by the wobbling of the shader effect. This effect means the player can have a large number of smoke clouds active without reducing the framerate to unplayable levels.