Power plant changes - more enemies, and easier to find all of them
One recurring issue was that in the warehouse section, players would go via one route only, and miss enemies on the other side. I placed some crates to block off one side (unless the player stacks some crates, which is fine because it rewards clever thinking), forcing the player to go around the other side and encounter the additional enemies.
I also added a few extra enemies near the end of the level.