WIP on new offhand attack - fire extinguisher
In addition to the larger fire extinguishers that can be picked up and thrown, this smaller one can be held and sprayed right in an enemy's face! Compared to the smoke grenades, it allows more convenient cloud creation, with the downside of reduced capacity and duration.

This is still very much a work in progress and there are bugs to iron out, such as the cloud not appearing properly the first time it's fired. There are also balance tweaks to deal with. But it's proving very promising so far!