GUI shader for filled radial menu slice
This method allows the size, shape, colour and outline to be changed by altering values of a material instance, and allows for much snazzier looking slices. The code is contained in a subgraph so it can be easily reused for different kinds of menu slices. For example, I'm making one that pulls double-duty as a resource meter.