Shader subgraph for fillable resource meter
This subgraph is for a resource bar that empties and fills. The 'drained' section of the graphic uses a scrolling animated Ben-Day dots texture for a nice comic book panel-feel, and the animation only appearing on the empty part acts as a growing indicator of low resources as the meter drains. The meter also shifts colour when dropping below a certain threshold. The subgraph uses a single 'length' input that can be calculated externally, allowing the meter to fill in any direction or pattern based on the shape of the desired graphic.