New GUI fonts
I've started using the free Antonio font, which I think carries the right balance of  sleek enough for the desired GUI visual style, and text-y enough for the comic book-style, while remaining easy to read. I may start using other fonts as well in other parts of the game, but for now this is a great replacement for the default Unity font.