DetectionProfile class to ensure consistent detection with particular attack types
This essentially takes a LayerMask and turns it into a scriptable object that can be referenced by multiple scripts, meaning that everything that needs to follow a particular detection rule will always use the same code. I find this especially important for things like enemy vision and attack detection, that can have complex relationships with different layers that need to be kept track of.
The current two I’ve implemented are ‘Standard Vision’ for enemy vision, and ‘Physical Attack’ for any attack that interacts with standard hard surfaces in the environment. I know I’ll need at least one more for when I add enemies with thermal vision.