WIP: smoke particles actually spreading
This complex system sorts particles into grid spaces, then checks each space for particles that are too close, and applies a small positional offset to make them move away from each other. This ensures that if a large number of particles are present, they actually billow out to form a larger cloud, in a much more realistic manner.
This system still needs improving, though. Not only is it dependent on a singleton class that has to be manually added to the scene, but a massive amount of garbage collection occurs with large numbers of particles, causing severe lag. But overall, this system is a massive improvement and I'm extremely proud of it so far.