SmokeParticleDensityController – reduced grid space checks per particle from 27 to 8
Checking grid spaces is supposed to reduce the number of particle checks for better optimisation, but the number of grid space checks turned out to be the real bottleneck. Initially it checked all adjacent spaces around the particle, but I realised that if each grid space is large enough (relative to the distance we want to check around each particle), the checking zone for each particle will only ever extend to an adjacent space on one side, rather than both. By calculating which direction that is (comparing the rounded grid space value to the non-rounded particle position in grid space), that means we only need to check a 2x2x2 cube instead of a 3x3x3 one.
This hasn’t 100% fixed performance, but it is still a big help and I’m proud of the design.