SmokeParticleDensityController – resolver vector now affects velocity rather than position
Previously, the resolver vectors would be directly applied as offsets to the particle positions. This is theoretically easier to keep track of each frame, but results in the particles heavily jittering and often clipping through surfaces.
I fixed this by instead modifying the velocity value, slowly shifting it towards the desired direction to ensure smoother and less jittery movement. This ensures the actual particle movement is always done through the ParticleSystem’s code, respecting collisions and preventing phasing through surfaces.